% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgstats.R
\name{get_cran_logs}
\alias{get_cran_logs}
\title{Download cran daily logs and builds data.table}
\usage{
get_cran_logs(from = as.Date("2021-03-02"), to = as.Date("2021-03-04"))
}
\arguments{
\item{from}{start date}

\item{to}{end date}
}
\value{
data.table
}
\description{
Downloads daily build logs (these are around 75MB per day so
don't try to download too many), combines them and returns.
A combined tmp file is stored under the directory as defined by the TMPDIR
environment variable that should be added to ~/.Renviron.
}
