\name{rpkgstats-package}
\alias{rpkgstats-package}
\alias{rpkgstats}
\docType{package}
\title{
R package usage statistics
}
\description{
Allow users to share download information of a package and rate and comment on the package within R directly. It requires the package RCurl if a user does not want to open a web browser.
}
\details{
\tabular{ll}{
Package: \tab rpkgstats\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-04-16\cr
License: \tab Free \cr
}
Allow users to share download information of a package and rate and comment on the package within R directly. The information is store on a webserver and can be viewed within R.
}
\author{
Zhiyong Zhang

Maintainer: Zhiyong Zhang <johnnyzhz@gmail.com>
}
