\name{uniqueID}
\alias{uniqueID}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Generate a unique (hopefully) id to identify a user
}
\description{
The unique information is only used to make sure a user won't rate a package multiple times within a short period of time. No private inforamtion can be obtained from such uqniue id. 
}
\usage{
uniqueID()
}
